+++ Changelog.md
@@ -5,6 +5,7 @@
 - Foundry 14: Don't let core Foundry expire turnEnd events.
 - `system.defenses.all.mod` will be added to all defenses (dodge, meleeWeapons, unarmedCombat, intimidation, maneuver, taunt, trick)
 - `system.defenses.physical.mod` will be added to dodge, meleeWeapons and unarmedCombat defenses.
+- `system.defenses.interaction.mod` will be added to intimidation, maneuver, taunt, trick
 
 ## 13.27.6 - damageMod for skills